Moving exported lodash replacements back to testing utils only
