--- .changeset/strange-days-refuse.md
@@ -1,5 +1,0 @@
----
-'@matrix-widget-toolkit/api': patch
----
-
-Adds uniqueId and deepEqual to exported utils due to lodash removal
